+++ src/python/pants/backend/go/build.py
@@ -8,7 +8,6 @@ from dataclasses import dataclass
 from pathlib import PurePath
 from typing import List, Optional
 
-from pants.backend.go.import_analysis import ResolvedImportPathsForGoLangDistribution
 from pants.backend.go.module import DownloadedExternalModule, DownloadExternalModuleRequest
 from pants.backend.go.pkg import (
     ResolvedGoPackage,
@@ -23,6 +22,7 @@ from pants.backend.go.target_types import (
     GoExternalModuleVersionField,
     GoPackageSources,
 )
+from pants.backend.go.util_rules.import_analysis import ResolvedImportPathsForGoLangDistribution
 from pants.backend.go.util_rules.sdk import GoSdkProcess
 from pants.build_graph.address import Address, AddressInput
 from pants.core.goals.package import (
