+++ src/python/pants/backend/go/target_type_rules.py
@@ -6,8 +6,7 @@ import logging
 from collections import defaultdict
 from dataclasses import dataclass
 
-from pants.backend.go import import_analysis, pkg
-from pants.backend.go.import_analysis import ResolvedImportPathsForGoLangDistribution
+from pants.backend.go import pkg
 from pants.backend.go.module import FindNearestGoModuleRequest, ResolvedOwningGoModule
 from pants.backend.go.pkg import (
     ResolvedGoPackage,
@@ -20,6 +19,8 @@ from pants.backend.go.target_types import (
     GoPackageDependencies,
     GoPackageSources,
 )
+from pants.backend.go.util_rules import import_analysis
+from pants.backend.go.util_rules.import_analysis import ResolvedImportPathsForGoLangDistribution
 from pants.base.specs import (
     AddressSpecs,
     DescendantAddresses,
