+++ src/python/pants/backend/go/util_rules/BUILD
@@ -2,3 +2,4 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_library()
+python_tests(name='tests')
